cut: remove single page extraction script, add multi-page extraction
The commit removes the bash script that was used to extract the first page from PDF files. In its place, a new script has been added which allows for extracting multiple pages from the start of PDF files. The number of pages to be extracted can be specified as an argument when running the script. If no argument is provided, it defaults to extracting just the first page.