update python version and refactor uv installation script
This commit updates the Python version in `.chezmoidata.yaml` to 3.12. It also refactors the `run_onchange_before_install-uv-python.sh` script by deleting it and replacing it with a templated version (`run_onchange_before_install-uv-python.sh.tmpl`). The new template uses the Python version specified in `.chezmoidata.yaml` as the default for uv tool installations.